all tests pass

+++ appointment-booking-api/tests/controllers/appointments/DELETE.test.ts
@@ -15,14 +15,16 @@ describe("DELETE request", () => {
   test("request to /api/appointments/:timestamp successfully deletes an appointment", async () => {
     const { appointment } = await createTestAppointment({ pushToDb: true });
 
+    if (appointment === null) throw Error();
+
     const response = await api.delete(
-      `/api/appointments/${appointment?.timestamp.toJSON()}`
+      `/api/appointments/${appointment.timestamp.toJSON()}`
     );
 
     expect(response.status).toBe(204);
 
     const deletedAppointment = await prisma.appointment.findUnique({
-      where: { timestamp: appointment?.timestamp },
+      where: { timestamp: appointment.timestamp },
     });
 
     expect(deletedAppointment).toBe(null);
